Added missing RealSwords plugins, all but one of Kuertee's mods, and a few misc changes.
The RealSwords stuff has been changed to all saying 'use FCOM_RealSwords' instead for FCOM builds, because even if the plugins are new and do not have an equivalent in FCOM_RealSwords, they should not be used with FCOM. Dev Akm is currently working on including these updates in the FCOM version anyway.

The kuertee mod that isn't in is Combat Timescale, as it is in an OMOD, and I couldn't open it on this machine, and the file name wasn't mentioned in the readme.

git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@63 bfee223e-1fc4-11de-b650-e715bd6d7cf1